requirements.txt
pandas
numpy
sqlalchemy
mysql-connector-python
PyYAML
matplotlib
plotly
streamlit
openpyxl